requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
pandas>=2.1.0
scikit-learn>=1.4.0
tqdm>=4.66.0